+++ R/modelspatial.R
@@ -1,29 +1,55 @@
 .modelspatial<-function(y, SpParams, meteo, model = "spwb", 
                         control = defaultControl(), dates = NULL, 
                         mergeTrees = FALSE,
-                        summaryFunction=NULL, args=NULL, progress = TRUE) {
+                        summaryFunction=NULL, args=NULL, 
+                        parallelize = FALSE, progress = TRUE) {
   sp = as(y,"SpatialPoints")
   topo = y@data
   spt = SpatialPointsTopography(sp, topo$elevation, topo$slope, topo$aspect)
   longlat = spTransform(sp,CRS("+proj=longlat"))
   latitude = longlat@coords[,2]
+  elevation = y@data$elevation
+  slope = y@data$slope
+  aspect = y@data$aspect
   
   control$verbose = FALSE
   
   forestlist = y@forestlist
   soillist  = y@soillist
-  xlist = y@xlist
+  xlist  = y@xlist
+  
   
   n = length(forestlist)
   reslist = vector("list",n)
+  
+  initf<-function(i) {
+    f = forestlist[[i]]
+    s = soillist[[i]]
+    x = NULL
+    if((!is.null(f)) && (!is.null(s))) {
+      if(mergeTrees) f = forest_mergeTrees(f)
+      if(model=="spwb") {
+        x = medfate::forest2spwbInput(f, s, SpParams, control)
+      } else if(model=="growth") {
+        x = medfate::forest2growthInput(f, s, SpParams, control)
+      }
+    }      
+    return(x)
+  }
+  
+  if(progress) cat("Initializing:\n")
   if(progress) pb = txtProgressBar(0, n, style=3)
   for(i in 1:n) {
     if(progress) setTxtProgressBar(pb, i)
+    xlist[[i]] = initf(i)
+  }
+  if(progress) cat("\n")
+  
+  simf<-function(i, sfun = NULL){
     f = forestlist[[i]]
     s = soillist[[i]]
+    x = xlist[[i]]
     if((!is.null(f)) && (!is.null(s))) {
-      if(mergeTrees) f = forest_mergeTrees(f)
-      # cat(" - Meteorology")
       if(inherits(meteo,"data.frame")) met = meteo
       else if(inherits(meteo,"SpatialPointsMeteorology") || inherits(meteo,"SpatialGridMeteorology")|| inherits(meteo,"SpatialPixelsMeteorology")) {
         met = meteo@data[[i]]
@@ -31,37 +57,57 @@
         met = meteoland::interpolationpoints(meteo, spt[i], dates=dates, verbose=FALSE)
         met = met@data[[1]]
       }
-      # cat(" - Soil water balance")
-      if(model=="spwb") {
-        xlist[[i]] = forest2spwbInput(f, s, SpParams, control)
-        S<-spwb(xlist[[i]], s, meteo=met, 
-                latitude = latitude[i], elevation = y@data$elevation[i],
-                slope = y@data$slope[i], aspect = y@data$aspect[i])      
-      } else if(model=="growth") {
-        xlist[[i]] = forest2growthInput(f, s, SpParams, control)
-        S<-growth(xlist[[i]], s, meteo=met, 
-                latitude = latitude[i], elevation = y@data$elevation[i],
-                slope = y@data$slope[i], aspect = y@data$aspect[i])      
-      }
       if(!is.null(dates)) met = met[as.character(dates),] #subset dates
-      if(!is.null(summaryFunction)){
-        #Fill argument list for  summary function
+      if(model=="spwb") {
+        S<-medfate::spwb(x, s, meteo=met, 
+                         latitude = latitude[i], elevation = elevation[i],
+                         slope = slope[i], aspect = aspect[i])      
+      } else if(model=="growth") {
+        S<-medfate::growth(x, s, meteo=met, 
+                           latitude = latitude[i], elevation = elevation[i],
+                           slope = slope[i], aspect = aspect[i])      
+      }
+      if(!is.null(sfun)){
         argList = list(object=S)
         if(!is.null(args)) {
           for(j in 1:length(args)) argList[names(args)[j]]=args[j]
         }
-        reslist[[i]] = do.call(summaryFunction, args=argList)
+        res = do.call(sfun, args=argList)
       } else {
-        reslist[[i]] = S
+        res = S
       }
+      return(res)
+    }
+    return(NULL)
+  }
+  
+  if(parallelize) {
+    if(progress) cat("Simulation...")
+    env<-environment()
+    cl<-makeCluster(detectCores()-1)
+    varlist = c("forestlist", "soillist", "mergeTrees", "meteo", "model",
+                "SpParams", "control", "latitude", "elevation", "slope", "aspect", 
+                "dates", "args", "xlist")
+    clusterExport(cl, varlist, envir = env)
+    reslist = clusterApply(cl, 1:n, simf, sfun = summaryFunction)
+    stopCluster(cl)
+    if(progress) cat("done.\n")
+  } else {
+    if(progress) cat("Simulation:\n")
+    if(progress) pb = txtProgressBar(0, n, style=3)
+    for(i in 1:n) {
+      if(progress) setTxtProgressBar(pb, i)
+      reslist[[i]] = simf(i, sfun = summaryFunction)
     }
   }
+
   return(list(xlist = xlist, reslist = reslist))
 }
 
 
 spwbpoints<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL, 
-                     summaryFunction=NULL, args=NULL, progress = TRUE) {
+                     summaryFunction=NULL, args=NULL, 
+                     parallelize = FALSE, progress = TRUE) {
   
   #Check input
   if(!inherits(y,"SpatialPointsLandscape")) 
@@ -77,13 +123,14 @@ spwbpoints<-function(y, SpParams, meteo, control = defaultControl(), dates = NUL
   }
 
   l = .modelspatial(y=y, SpParams = SpParams, meteo = meteo, model = "spwb", control = control, dates = dates, 
-                    summaryFunction = summaryFunction, args = args, progress = progress)
+                    summaryFunction = summaryFunction, args = args, parallelize = parallelize, progress = progress)
   res = list(coords = y@coords, bbox = y@bbox, proj4string = y@proj4string, xlist = l$xlist, reslist = l$reslist)
   class(res) = c("spwbpoints","list")
   return(res)
 }
 spwbgrid<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL, 
-                   summaryFunction=NULL, args=NULL, progress = TRUE) {
+                   summaryFunction=NULL, args=NULL, 
+                   parallelize = FALSE, progress = TRUE) {
   
   #Check input
   if(!inherits(y,"SpatialGridLandscape")) 
@@ -99,13 +146,14 @@ spwbgrid<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL,
   }
   
   l = .modelspatial(y=y, SpParams = SpParams, meteo = meteo, model = "spwb", control = control, dates = dates, 
-                    summaryFunction = summaryFunction, args = args, progress = progress)
+                    summaryFunction = summaryFunction, args = args, parallelize = parallelize, progress = progress)
   res = list(grid = y@grid, bbox = y@bbox, proj4string = y@proj4string, xlist = l$xlist, reslist = l$reslist)
   class(res) = c("spwbgrid","list")
   return(res)
 }
-spwbpixels<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL, summaryFunction=NULL, 
-                     args=NULL, progress = TRUE) {
+spwbpixels<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL, 
+                     summaryFunction=NULL, args=NULL, 
+                     parallelize = FALSE, progress = TRUE) {
   
   #Check input
   if(!inherits(y,"SpatialPixelsLandscape")) 
@@ -123,13 +171,14 @@ spwbpixels<-function(y, SpParams, meteo, control = defaultControl(), dates = NUL
   }
   
   l = .modelspatial(y=y, SpParams = SpParams, meteo = meteo, model = "spwb", control = control, dates = dates, 
-                    summaryFunction = summaryFunction, args = args, progress = progress)
+                    summaryFunction = summaryFunction, args = args, parallelize = parallelize, progress = progress)
   res = list(coords = y@coords, coords.nrs = y@coords.nrs, grid = y@grid, grid.index = y@grid.index, bbox = y@bbox, proj4string = y@proj4string, xlist = l$xlist, reslist = l$reslist)
   class(res) = c("spwbpixels","list")
   return(res)
 }
 growthpoints<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL, 
-                       summaryFunction=NULL, args=NULL, progress = TRUE) {
+                       summaryFunction=NULL, args=NULL, 
+                       parallelize = FALSE, progress = TRUE) {
   
   #Check input
   if(!inherits(y,"SpatialPointsLandscape")) 
@@ -145,13 +194,14 @@ growthpoints<-function(y, SpParams, meteo, control = defaultControl(), dates = N
   }
   
   l = .modelspatial(y=y, SpParams = SpParams, meteo = meteo, model = "growth", control = control, dates = dates, 
-                    summaryFunction = summaryFunction, args = args, progress = progress)
+                    summaryFunction = summaryFunction, args = args, parallelize = parallelize, progress = progress)
   res = list(coords = y@coords, bbox = y@bbox, proj4string = y@proj4string, xlist = l$xlist, reslist = l$reslist)
   class(res) = c("growthpoints","list")
   return(res)
 }
 growthgrid<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL, 
-                     summaryFunction=NULL, args=NULL, progress = TRUE) {
+                     summaryFunction=NULL, args=NULL, 
+                     parallelize = FALSE, progress = TRUE) {
   
   #Check input
   if(!inherits(y,"SpatialGridLandscape")) 
@@ -167,13 +217,14 @@ growthgrid<-function(y, SpParams, meteo, control = defaultControl(), dates = NUL
   }
   
   l = .modelspatial(y=y, SpParams = SpParams, meteo = meteo, model = "growth", control = control, dates = dates, 
-                    summaryFunction = summaryFunction, args = args, progress = progress)
+                    summaryFunction = summaryFunction, args = args, parallelize = parallelize, progress = progress)
   res = list(grid = y@grid, bbox = y@bbox, proj4string = y@proj4string, xlist = l$xlist, reslist = l$reslist)
   class(res) = c("growthgrid","list")
   return(res)
 }
 growthpixels<-function(y, SpParams, meteo, control = defaultControl(), dates = NULL, 
-                       summaryFunction=NULL, args=NULL, progress = TRUE) {
+                       summaryFunction=NULL, args=NULL, 
+                       parallelize = FALSE, progress = TRUE) {
   
   #Check input
   if(!inherits(y,"SpatialPixelsLandscape")) 
@@ -191,7 +242,7 @@ growthpixels<-function(y, SpParams, meteo, control = defaultControl(), dates = N
   }
   
   l = .modelspatial(y=y, SpParams = SpParams, meteo = meteo, model = "growth", control = control, dates = dates, 
-                    summaryFunction = summaryFunction, args = args, progress = progress)
+                    summaryFunction = summaryFunction, args = args, parallelize = parallelize, progress = progress)
   res = list(coords = y@coords, coords.nrs = y@coords.nrs, grid = y@grid, grid.index = y@grid.index, bbox = y@bbox, proj4string = y@proj4string, xlist = l$xlist, reslist = l$reslist)
   class(res) = c("growthpixels","list")
   return(res)