+++ data-raw/example_data.R
@@ -1,5 +1,5 @@
 #Example watershed
-examplewatershed = readRDS(paste0("D:/Datasets/Hydrology/Products/DistributedWatersheds/CanVila_watershed_30m.rds")) 
+examplewatershed = readRDS(paste0("D:/Datasets/Hydrology/Products/DistributedWatersheds/CanVila_watershed_100m.rds")) 
 usethis::use_data(examplewatershed, overwrite = T)
 
 #Example SpatialPointsLandscape